+++ Tests/Integration/MSFT_ADComputer.Integration.Tests.ps1
@@ -93,7 +93,7 @@ try
             }
 
             It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -BeTrue
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }
 
@@ -140,7 +140,7 @@ try
             }
 
             It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -BeTrue
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }
 
@@ -187,7 +187,7 @@ try
             }
 
             It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -BeTrue
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }
 
@@ -234,7 +234,7 @@ try
             }
 
             It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -BeTrue
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }
 
@@ -285,7 +285,7 @@ try
             }
 
             It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -BeTrue
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }
 
@@ -349,7 +349,7 @@ try
             }
 
             It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -BeTrue
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }
 
@@ -380,7 +380,7 @@ try
             }
 
             It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -BeTrue
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }
     }