+++ Tests/Integration/MSFT_ADDomainTrust.Integration.Tests.ps1
@@ -81,7 +81,7 @@ try
             }
 
             It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -BeTrue
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }
 
@@ -132,7 +132,7 @@ try
             }
 
             It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -BeTrue
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }
 
@@ -183,7 +183,7 @@ try
             }
 
             It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -BeTrue
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }
 
@@ -234,7 +234,7 @@ try
             }
 
             It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -BeTrue
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }
 
@@ -285,7 +285,7 @@ try
             }
 
             It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -BeTrue
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }
 
@@ -336,7 +336,7 @@ try
             }
 
             It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -BeTrue
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }
     }