+++ Tests/Integration/MSFT_ADUser.Integration.Tests.ps1
@@ -80,7 +80,7 @@ try
             }
 
             It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -BeTrue
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }
 
@@ -130,7 +130,7 @@ try
             }
 
             It 'Should return $true when Test-DscConfiguration is run' {
-                Test-DscConfiguration -Verbose | Should -BeTrue
+                Test-DscConfiguration -Verbose | Should -Be 'True'
             }
         }
     }