+++ CHANGELOG.md
@@ -57,6 +57,8 @@
     Azure Automation ([issue #416](https://github.com/PowerShell/ActiveDirectoryDsc/issues/416)).
   - Fix examples that had duplicate guid that would have prevented them
     to be published.
+  - Integration tests are now correctly evaluates the value from `Test-DscConfiguration`
+    ([issue #434](https://github.com/PowerShell/ActiveDirectoryDsc/issues/434)).
 - Changes to ADManagedServiceAccount
   - Added a requirement to README stating "Group Managed Service Accounts
     need at least one Windows Server 2012 Domain Controller"